Working in progress

+++ contracts/nft/SAToken.sol
@@ -140,7 +140,7 @@ contract SAToken is ISAToken, SATokenData, ERC721, ERC721Enumerable, LevelAccess
 
   function getBundle(uint256 tokenId) external view override returns (SA[] memory) {
     SA[] memory sas = new SA[](_bundles[tokenId].length);
-    for (uint i=0;i< _bundles[tokenId].length; i++) {
+    for (uint256 i = 0; i < _bundles[tokenId].length; i++) {
       sas[i] = _unpackUint256(_bundles[tokenId][i]);
     }
     return sas;
